sync-tool/requirements.txt
pymongo>=4.6.0
python-dotenv>=1.0.0
Pillow>=10.1.0
